refactor: move bottom sheet to note list
